Removidos os raise de exceções da camada de repository e algumas alterações

+++ gerenciamento_hexagonal/application/services.py
@@ -53,25 +53,30 @@ class GerenciamentoPropostaServices:
         return gerenciamento_proposta
 
     async def update_gerenciamento_proposta(self, gerenciamento_proposta_id: int, gerenciamento_proposta: GerenciamentoPropostaDTO) -> GerenciamentoProposta:
+        await self.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
+
         gerenciamento_proposta = GerenciamentoProposta(**gerenciamento_proposta.model_dump())
         gerenciamento_proposta = await self.repository.update_gerenciamento_proposta(gerenciamento_proposta_id, gerenciamento_proposta)
-        if not gerenciamento_proposta:
-            raise NotFoundError(f'gerenciamento_proposta with ID {gerenciamento_proposta_id} not found')
 
         return gerenciamento_proposta
 
     async def delete_gerenciamento_proposta(self, gerenciamento_proposta_id: int):
-        delete = await self.repository.delete_gerenciamento_proposta(gerenciamento_proposta_id)
-        if not delete:
-            raise NotFoundError(f'gerenciamento_proposta with ID {gerenciamento_proposta_id} not found')
+        await self.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
 
-        return delete
+        NotNullViolation = await self.repository.checar_gerenciamento_caracterizacao_exists(gerenciamento_proposta_id)
+
+        if NotNullViolation:
+            raise NotNullViolationError('cannot delete a gerenciamento_proposta when there is an associated gerenciamento_caracterizacao')
+
+        await self.repository.delete_gerenciamento_proposta(gerenciamento_proposta_id)
+
+        return {'message': f'gerenciamento_proposta with ID {gerenciamento_proposta_id} deleted'}
 
     async def create_gerenciamento_proposta_comentario(self, gerenciamento_proposta_id: int, gerenciamento_comentario: GerenciamentoComentarioDTO) -> GerenciamentoProposta:
+        await self.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
+
         gerenciamento_comentario = GerenciamentoComentario(**gerenciamento_comentario.model_dump())
         gerenciamento_proposta = await self.repository.create_gerenciamento_proposta_comentario(gerenciamento_proposta_id, gerenciamento_comentario)
-        if not gerenciamento_proposta:
-            raise NotFoundError(f'gerenciamento_proposta with ID {gerenciamento_proposta_id} not found')
 
         return gerenciamento_proposta
 
@@ -87,31 +92,32 @@ class GerenciamentoComentarioServices:
 
     async def get_gerenciamento_comentario_by_id(self, gerenciamento_comentario_id: int) -> GerenciamentoComentario:
         gerenciamento_comentario = await self.repository.get_gerenciamento_comentario_by_id(gerenciamento_comentario_id)
+
         if not gerenciamento_comentario:
             raise NotFoundError(f'gerenciamento_comentario with ID {gerenciamento_comentario_id} not found')
 
         return gerenciamento_comentario
 
     async def update_gerenciamento_comentario(self, gerenciamento_comentario_id: int, gerenciamento_comentario: GerenciamentoComentarioDTO) -> GerenciamentoComentario:
+        await self.get_gerenciamento_comentario_by_id(gerenciamento_comentario_id)
+
         gerenciamento_comentario = GerenciamentoComentario(**gerenciamento_comentario.model_dump())
         gerenciamento_comentario = await self.repository.update_gerenciamento_comentario(gerenciamento_comentario_id, gerenciamento_comentario)
-        if not gerenciamento_comentario:
-            raise NotFoundError(f'gerenciamento_comentario with ID {gerenciamento_comentario_id} not found')
 
         return gerenciamento_comentario
 
     async def delete_gerenciamento_comentario(self, gerenciamento_comentario_id: int) -> None:
-        delete = await self.repository.delete_gerenciamento_comentario(gerenciamento_comentario_id)
+        await self.get_gerenciamento_comentario_by_id(gerenciamento_comentario_id)
 
-        if not delete:
-            raise NotFoundError(f'gerenciamento_comentario with ID {gerenciamento_comentario_id} not found')
+        await self.repository.delete_gerenciamento_comentario(gerenciamento_comentario_id)
 
-        return delete
+        return {'message': 'gerenciamento_comentario with ID {gerenciamento_comentario_id} deleted'}
 
 
 class GerenciamentoMetaServices:
-    def __init__(self, repository_gerenciamento: GerenciamentoMetaRepository):
+    def __init__(self, repository_gerenciamento: GerenciamentoMetaRepository, service_proposta: GerenciamentoPropostaServices):
         self.repository = repository_gerenciamento
+        self.service_proposta = service_proposta
 
     async def get_gerenciamento_meta(self) -> list[GerenciamentoMeta]:
         gerenciamento_metas = await self.repository.get_gerenciamento_meta()
@@ -126,29 +132,27 @@ class GerenciamentoMetaServices:
         return gerenciamento_meta
 
     async def create_gerenciamento_meta(self, gerenciamento_proposta_id: int, gerenciamento_meta: GerenciamentoMetaDTO) -> GerenciamentoMeta:
-        try:
-            gerenciamento_meta = GerenciamentoMeta(**gerenciamento_meta.model_dump())
-            gerenciamento_meta = await self.repository.create_gerenciamento_meta(gerenciamento_proposta_id, gerenciamento_meta)
+        await self.service_proposta.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
 
-            return gerenciamento_meta
+        gerenciamento_meta = GerenciamentoMeta(**gerenciamento_meta.model_dump())
+        gerenciamento_meta = await self.repository.create_gerenciamento_meta(gerenciamento_proposta_id, gerenciamento_meta)
 
-        except NotFoundError as e:
-            raise e
+        return gerenciamento_meta
 
     async def update_gerenciamento_meta(self, gerenciamento_meta_id: int, gerenciamento_meta: GerenciamentoMetaDTO) -> GerenciamentoMeta:
+        await self.get_gerenciamento_meta_by_id(gerenciamento_meta_id)
+
         gerenciamento_meta = GerenciamentoMeta(**gerenciamento_meta.model_dump())
         gerenciamento_meta = await self.repository.update_gerenciamento_meta(gerenciamento_meta_id, gerenciamento_meta)
-        if not gerenciamento_meta:
-            raise NotFoundError(f'gerenciamento_meta with ID {gerenciamento_meta_id} not found')
 
         return gerenciamento_meta
 
     async def delete_gerenciamento_meta(self, gerenciamento_meta_id: int) -> bool:
-        delete = await self.repository.delete_gerenciamento_meta(gerenciamento_meta_id)
-        if not delete:
-            raise NotFoundError(f'gerenciamento_meta with ID {gerenciamento_meta_id} not found')
+        await self.get_gerenciamento_meta_by_id(gerenciamento_meta_id)
 
-        return True
+        await self.repository.delete_gerenciamento_meta(gerenciamento_meta_id)
+
+        return {'message': f'gerenciamento_meta with ID {gerenciamento_meta_id} deleted'}
 
 
 class GerenciamentoQuantitativoServices:
@@ -172,7 +176,7 @@ class GerenciamentoQuantitativoServices:
     async def create_gerenciamento_quantitativo(self, gerenciamento_proposta_id: int, gerenciamento_quantitativo: GerenciamentoQuantitativoDTO) -> GerenciamentoQuantitativo:
         await self.service_proposta.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
 
-        UniqueViolationCheck = await self.repository.find_by_gerenciamento_proposta_id(gerenciamento_proposta_id)
+        UniqueViolationCheck = await self.repository.checar_associacao_existente(gerenciamento_proposta_id)
 
         if UniqueViolationCheck:
             raise UniqueViolation(f'gerenciamento_quantitativo already exist for gerenciamento_proposta ID: {gerenciamento_proposta_id}')
@@ -193,7 +197,7 @@ class GerenciamentoQuantitativoServices:
     async def delete_gerenciamento_quantitativo(self, gerenciamento_quantitativo_id: int):
         await self.get_gerenciamento_quantitativo_by_id(gerenciamento_quantitativo_id)
 
-        NotNullViolation = await self.repository.check_gerenciamento_caracterizacao_exist(gerenciamento_quantitativo_id)
+        NotNullViolation = await self.repository.checar_gerenciamento_caracterizacao_exists(gerenciamento_quantitativo_id)
 
         if NotNullViolation:
             raise NotNullViolationError('cannot delete a gerenciamento_quantitativo when there is an associated gerenciamento_caracterizacao')
@@ -212,8 +216,9 @@ class GerenciamentoQuantitativoServices:
 
 
 class GerenciamentoQualitativoServices:
-    def __init__(self, repository_gerenciamento: GerenciamentoQualitativoRepository):
+    def __init__(self, repository_gerenciamento: GerenciamentoQualitativoRepository, service_proposta: GerenciamentoPropostaServices):
         self.repository = repository_gerenciamento
+        self.service_proposta = service_proposta
 
     async def get_gerenciamento_qualitativo(self) -> list[GerenciamentoQualitativo]:
         gerenciamento_qualitativos = await self.repository.get_gerenciamento_qualitativo()
@@ -221,55 +226,48 @@ class GerenciamentoQualitativoServices:
         return gerenciamento_qualitativos
 
     async def get_gerenciamento_qualitativo_by_id(self, gerenciamento_qualitativo_id: int) -> GerenciamentoQualitativo:
-        try:
-            gerenciamento_qualitativo = await self.repository.get_gerenciamento_qualitativo_by_id(gerenciamento_qualitativo_id)
+        gerenciamento_qualitativo = await self.repository.get_gerenciamento_qualitativo_by_id(gerenciamento_qualitativo_id)
 
-            return gerenciamento_qualitativo
-
-        except NotFoundError:
+        if not gerenciamento_qualitativo:
             raise NotFoundError(f'gerenciamento_qualitativo with ID: {gerenciamento_qualitativo_id} not found')
 
+        return gerenciamento_qualitativo
+
     async def create_gerenciamento_qualitativo(self, gerenciamento_proposta_id: int, gerenciamento_qualitativo: GerenciamentoQualitativoDTO) -> GerenciamentoQualitativo:
-        try:
-            gerenciamento_qualitativo = GerenciamentoQualitativo(**gerenciamento_qualitativo.model_dump())
-            gerenciamento_qualitativo = await self.repository.create_gerenciamento_qualitativo(gerenciamento_proposta_id, gerenciamento_qualitativo)
+        await self.service_proposta.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
 
-            return gerenciamento_qualitativo
+        UniqueViolationCheck = await self.repository.checar_associacao_existente(gerenciamento_proposta_id)
 
-        except NotFoundError:
-            raise NotFoundError(f'gerenciamento_proposta with ID: {gerenciamento_proposta_id} not found')
-
-        except UniqueViolation:
+        if UniqueViolationCheck:
             raise UniqueViolation(f'gerenciamento_qualitativo already exist for gerenciamento_proposta ID: {gerenciamento_proposta_id}')
 
+        gerenciamento_qualitativo = GerenciamentoQualitativo(**gerenciamento_qualitativo.model_dump())
+        gerenciamento_qualitativo = await self.repository.create_gerenciamento_qualitativo(gerenciamento_proposta_id, gerenciamento_qualitativo)
+
+        return gerenciamento_qualitativo
+
     async def update_gerenciamento_qualitativo(self, gerenciamento_qualitativo_id: int, gerenciamento_qualitativo: GerenciamentoQualitativoDTO) -> GerenciamentoQualitativo | None:
-        try:
-            gerenciamento_qualitativo = GerenciamentoQualitativo(**gerenciamento_qualitativo.model_dump())
-            gerenciamento_qualitativo = await self.repository.update_gerenciamento_qualitativo(gerenciamento_qualitativo_id, gerenciamento_qualitativo)
+        await self.get_gerenciamento_qualitativo_by_id(gerenciamento_qualitativo_id)
 
-            return gerenciamento_qualitativo
+        gerenciamento_qualitativo = GerenciamentoQualitativo(**gerenciamento_qualitativo.model_dump())
+        gerenciamento_qualitativo = await self.repository.update_gerenciamento_qualitativo(gerenciamento_qualitativo_id, gerenciamento_qualitativo)
 
-        except NotFoundError as e:
-            raise e
+        return gerenciamento_qualitativo
 
     async def delete_gerenciamento_qualitativo(self, gerenciamento_qualitativo_id: int):
-        try:
-            result = await self.repository.delete_gerenciamento_qualitativo(gerenciamento_qualitativo_id)
+        await self.get_gerenciamento_qualitativo_by_id(gerenciamento_qualitativo_id)
 
-            return result
+        await self.repository.delete_gerenciamento_qualitativo(gerenciamento_qualitativo_id)
 
-        except NotFoundError as e:
-            raise e
+        return {'message': f'gerenciamento_qualitativo with ID {gerenciamento_qualitativo_id} deleted'}
 
     async def create_gerenciamento_qualitativo_comentario(self, gerenciamento_qualitativo_id: int, gerenciamento_comentario: GerenciamentoComentarioDTO) -> GerenciamentoQualitativo:
-        try:
-            gerenciamento_comentario = GerenciamentoComentario(**gerenciamento_comentario.model_dump())
-            gerenciamento_qualitativo_comentario = await self.repository.create_gerenciamento_qualitativo_comentario(gerenciamento_qualitativo_id, gerenciamento_comentario)
+        await self.repository.get_gerenciamento_qualitativo_by_id(gerenciamento_qualitativo_id)
 
-            return gerenciamento_qualitativo_comentario
+        gerenciamento_comentario = GerenciamentoComentario(**gerenciamento_comentario.model_dump())
+        gerenciamento_qualitativo_comentario = await self.repository.create_gerenciamento_qualitativo_comentario(gerenciamento_qualitativo_id, gerenciamento_comentario)
 
-        except NotFoundError as e:
-            raise e
+        return gerenciamento_qualitativo_comentario
 
 
 class GerenciamentoCaracterizacaoServices:
@@ -293,7 +291,7 @@ class GerenciamentoCaracterizacaoServices:
     async def create_gerenciamento_caracterizacao(self, gerenciamento_quantitativo_id: int, gerenciamento_caracterizacao: GerenciamentoCaracterizacaoDTO) -> GerenciamentoCaracterizacao:
         await self.service_quantitativo.get_gerenciamento_quantitativo_by_id(gerenciamento_quantitativo_id)
 
-        UniqueViolationCheck = await self.repository.find_by_gerenciamento_quantitativo_id(gerenciamento_quantitativo_id)
+        UniqueViolationCheck = await self.repository.checar_associacao_existente(gerenciamento_quantitativo_id)
 
         if UniqueViolationCheck:
             raise UniqueViolation(f'gerenciamento_caracterizacao already exist for gerenciamento_quantitativo ID: {gerenciamento_quantitativo_id}')
